+++ src/Header/CacheControl.php
@@ -4,7 +4,7 @@ declare(strict_types = 1);
 namespace Innmind\Http\Header;
 
 use Innmind\Http\Header as HeaderInterface;
-use Innmind\Immutable\Set;
+use Innmind\Immutable\Sequence;
 
 /**
  * @psalm-immutable
@@ -25,7 +25,7 @@ final class CacheControl implements HeaderInterface
     }
 
     #[\Override]
-    public function values(): Set
+    public function values(): Sequence
     {
         return $this->header->values();
     }