+++ src/Header/Cookie.php
@@ -5,7 +5,7 @@ namespace Innmind\Http\Header;
 
 use Innmind\Http\Header as HeaderInterface;
 use Innmind\Immutable\{
-    Set,
+    Sequence,
     Map,
 };
 
@@ -39,7 +39,7 @@ final class Cookie implements HeaderInterface
     }
 
     #[\Override]
-    public function values(): Set
+    public function values(): Sequence
     {
         return $this->header->values();
     }