+++ src/Header/IfUnmodifiedSince.php
@@ -5,7 +5,7 @@ namespace Innmind\Http\Header;
 
 use Innmind\Http\Header as HeaderInterface;
 use Innmind\TimeContinuum\PointInTime;
-use Innmind\Immutable\Set;
+use Innmind\Immutable\Sequence;
 
 /**
  * @psalm-immutable
@@ -36,7 +36,7 @@ final class IfUnmodifiedSince implements HeaderInterface
     }
 
     #[\Override]
-    public function values(): Set
+    public function values(): Sequence
     {
         return $this->header->values();
     }