+++ tests/Header/ContentRangeTest.php
@@ -8,7 +8,7 @@ use Innmind\Http\{
     Header,
     Header\ContentRangeValue
 };
-use Innmind\Immutable\Set;
+use Innmind\Immutable\Sequence;
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
 class ContentRangeTest extends TestCase
@@ -22,7 +22,7 @@ class ContentRangeTest extends TestCase
         $this->assertInstanceOf(Header::class, $h);
         $this->assertSame('Content-Range', $h->name());
         $v = $h->values();
-        $this->assertInstanceOf(Set::class, $v);
+        $this->assertInstanceOf(Sequence::class, $v);
         $this->assertSame($cr, $v->find(static fn() => true)->match(
             static fn($first) => $first,
             static fn() => null,